Extract roadmap dependency analysis and window resolution into dedicated services, improve variable scoping in charts, and refactor `RoadmapService`.
