+++ app/Services/Projects/RoadmapService.php
@@ -3,8 +3,6 @@
 namespace App\Services\Projects;
 
 use App\Models\Issue;
-use App\Models\IssueStatus;
-use App\Models\IssueStatusEvent;
 use App\Models\Milestone;
 use App\Models\Project;
 use Illuminate\Support\Carbon;
@@ -13,6 +11,12 @@ use Illuminate\Support\Str;
 
 final class RoadmapService
 {
+    public function __construct(
+        private readonly RoadmapBurnChartBuilder $burnChartBuilder,
+        private readonly RoadmapDependencyAnalyzer $dependencyAnalyzer,
+        private readonly RoadmapWindowResolver $windowResolver,
+    ) {}
+
     /**
      * @return array{
      *     burnChart: array<string, mixed>,
@@ -29,7 +33,56 @@ final class RoadmapService
     {
         $groupBy = in_array($groupBy, ['milestone', 'parent'], true) ? $groupBy : 'milestone';
 
-        $milestones = $project->milestones()
+        $milestones = $this->loadMilestones($project);
+        $selectedMilestoneId = $this->resolveMilestoneSelection($milestones, $selectedMilestoneId);
+
+        $issues = $this->loadRoadmapIssues($project);
+        $issuesById = $issues->keyBy(static fn (Issue $issue): string => (string) $issue->getKey());
+        $rootScopeIds = $this->resolveRootScopeIds($issues, $issuesById);
+
+        [$groups, $groupIdByIssueId] = $groupBy === 'parent'
+            ? $this->buildParentGroups($project, $issues, $issuesById, $rootScopeIds)
+            : $this->buildMilestoneGroups($project, $milestones, $issues, $rootScopeIds);
+
+        [
+            'dependencyEdges' => $dependencyEdges,
+            'groups' => $groups,
+        ] = $this->dependencyAnalyzer->analyze(
+            project: $project,
+            groups: $groups,
+            groupIdByIssueId: $groupIdByIssueId,
+            issues: $issues,
+            groupBy: $groupBy,
+        );
+
+        $groups = $this->finalizeGroups($groups, $issuesById);
+        $summary = $this->buildSummary($milestones, $issues, $rootScopeIds, $groups);
+        $timelineChart = $this->buildTimelineChart($groups);
+        $milestoneOptions = $this->buildMilestoneOptions($milestones, $issues);
+
+        $burnChart = $this->burnChartBuilder->build(
+            milestone: $milestones->firstWhere('id', $selectedMilestoneId),
+            issues: $this->loadBurnChartIssues($selectedMilestoneId),
+        );
+
+        return [
+            'burnChart' => $burnChart,
+            'dependencyEdges' => $dependencyEdges,
+            'groupBy' => $groupBy,
+            'groups' => $groups,
+            'milestoneOptions' => $milestoneOptions,
+            'selectedMilestoneId' => $selectedMilestoneId,
+            'summary' => $summary,
+            'timelineChart' => $timelineChart,
+        ];
+    }
+
+    /**
+     * @return Collection<int, Milestone>
+     */
+    private function loadMilestones(Project $project): Collection
+    {
+        return $project->milestones()
             ->orderByRaw('CASE WHEN due_at IS NULL THEN 1 ELSE 0 END')
             ->orderBy('due_at')
             ->orderByRaw('CASE WHEN starts_at IS NULL THEN 1 ELSE 0 END')
@@ -46,22 +99,25 @@ final class RoadmapService
                 'due_at',
                 'released_at',
             ]);
+    }
 
-        $issues = Issue::query()
+    /**
+     * @return Collection<int, Issue>
+     */
+    private function loadRoadmapIssues(Project $project): Collection
+    {
+        return Issue::query()
             ->where('project_id', $project->id)
             ->with([
                 'status:id,name,color,is_done',
                 'type:id,name,key,tier',
-                'assignee:id,name',
                 'milestone:id,project_id,name,type,state,starts_at,due_at,released_at,version',
-                'project:id,key,name',
                 'incomingLinks' => fn ($query) => $query
                     ->whereHas('type', fn ($typeQuery) => $typeQuery->where('key', 'blocks'))
                     ->with([
                         'type:id,key,name,inverse_name,is_symmetric',
-                        'from:id,key,summary,project_id,issue_status_id,milestone_id,parent_id,issue_type_id,children_count,progress_percent,starts_at,due_at,assignee_id,created_at,updated_at',
+                        'from:id,key,summary,project_id,issue_status_id,milestone_id',
                         'from.status:id,name,color,is_done',
-                        'from.type:id,name,key,tier',
                         'from.milestone:id,project_id,name,type,state,starts_at,due_at,released_at,version',
                         'from.project:id,key,name',
                     ]),
@@ -78,49 +134,33 @@ final class RoadmapService
                 'assignee_id',
                 'story_points',
                 'children_count',
-                'progress_percent',
                 'starts_at',
                 'due_at',
                 'created_at',
                 'updated_at',
             ]);
+    }
 
-        $issuesById = $issues->keyBy(static fn (Issue $issue): string => (string) $issue->getKey());
-        $rootScopeIds = $this->resolveRootScopeIds($issues, $issuesById);
+    /**
+     * @return Collection<int, Issue>
+     */
+    private function loadBurnChartIssues(?string $selectedMilestoneId): Collection
+    {
+        if ($selectedMilestoneId === null) {
+            return collect();
+        }
 
-        [$groups, $groupIdByIssueId] = $groupBy === 'parent'
-            ? $this->buildParentGroups($project, $issues, $issuesById, $rootScopeIds)
-            : $this->buildMilestoneGroups($project, $milestones, $issues, $rootScopeIds);
-
-        [$groups, $dependencyEdges] = $this->applyDependencyRisk(
-            $project,
-            $groups,
-            $groupIdByIssueId,
-            $issues,
-            $issuesById,
-            $groupBy
-        );
-
-        $groups = $this->finalizeGroups($groups, $issuesById);
-        $summary = $this->buildSummary($milestones, $issues, $rootScopeIds, $groups);
-        $timelineChart = $this->buildTimelineChart($groups);
-        $milestoneOptions = $this->buildMilestoneOptions($milestones, $issues);
-        $selectedMilestoneId = $this->resolveMilestoneSelection($milestones, $milestoneOptions, $selectedMilestoneId);
-        $burnChart = $this->buildBurnChart(
-            $milestones->firstWhere('id', $selectedMilestoneId),
-            $issues->where('milestone_id', $selectedMilestoneId)->values()
-        );
-
-        return [
-            'burnChart' => $burnChart,
-            'dependencyEdges' => $dependencyEdges,
-            'groupBy' => $groupBy,
-            'groups' => $groups,
-            'milestoneOptions' => $milestoneOptions,
-            'selectedMilestoneId' => $selectedMilestoneId,
-            'summary' => $summary,
-            'timelineChart' => $timelineChart,
-        ];
+        return Issue::query()
+            ->where('milestone_id', $selectedMilestoneId)
+            ->with('status:id,name,color,is_done')
+            ->get([
+                'id',
+                'key',
+                'milestone_id',
+                'issue_status_id',
+                'created_at',
+                'updated_at',
+            ]);
     }
 
     /**
@@ -198,8 +238,8 @@ final class RoadmapService
                 'subtitle' => $milestone->version ?: null,
                 'state_label' => Str::headline((string) ($milestone->state?->value ?? $milestone->state ?? 'planned')),
                 'url' => route('projects.milestones.show', [$project, $milestone]),
-                'starts_at' => $this->normalizeDate($milestone->starts_at),
-                'ends_at' => $this->normalizeDate($milestone->released_at ?? $milestone->due_at),
+                'starts_at' => $this->windowResolver->normalizeDate($milestone->starts_at),
+                'ends_at' => $this->windowResolver->normalizeDate($milestone->released_at ?? $milestone->due_at),
                 '_issue_ids' => [],
                 '_scope_ids' => [],
                 '_milestone_ids' => [$milestone->id => true],
@@ -280,8 +320,8 @@ final class RoadmapService
                         'subtitle' => $rootIssue->key,
                         'state_label' => $rootIssue->status?->name,
                         'url' => route('issues.show', ['project' => $project, 'issue' => $rootIssue]),
-                        'starts_at' => $this->normalizeDate($rootIssue->starts_at),
-                        'ends_at' => $this->normalizeDate($rootIssue->due_at),
+                        'starts_at' => $this->windowResolver->normalizeDate($rootIssue->starts_at),
+                        'ends_at' => $this->windowResolver->normalizeDate($rootIssue->due_at),
                         '_issue_ids' => [],
                         '_scope_ids' => [$rootId => true],
                         '_milestone_ids' => [],
@@ -329,161 +369,6 @@ final class RoadmapService
 
     /**
      * @param  array<string, array<string, mixed>>  $groups
-     * @param  array<string, string>  $groupIdByIssueId
-     * @param  Collection<int, Issue>  $issues
-     * @param  Collection<string, Issue>  $issuesById
-     * @return array{0: array<string, array<string, mixed>>, 1: array<int, array<string, mixed>>}
-     */
-    private function applyDependencyRisk(
-        Project $project,
-        array $groups,
-        array $groupIdByIssueId,
-        Collection $issues,
-        Collection $issuesById,
-        string $groupBy
-    ): array {
-        $dependencyEdges = [];
-
-        foreach ($issues as $issue) {
-            if ($issue->status?->is_done) {
-                continue;
-            }
-
-            $issueId = (string) $issue->getKey();
-            $targetGroupId = $groupIdByIssueId[$issueId] ?? null;
-            if ($targetGroupId === null || ! isset($groups[$targetGroupId])) {
-                continue;
-            }
-
-            $activeBlockers = $issue->incomingLinks
-                ->filter(static fn ($link) => $link->from && ! ($link->from->status?->is_done ?? false))
-                ->values();
-
-            if ($activeBlockers->isEmpty()) {
-                continue;
-            }
-
-            $groups[$targetGroupId]['_blocked_issue_ids'][$issueId] = true;
-
-            if (count($groups[$targetGroupId]['blocked_issue_samples']) < 3) {
-                $groups[$targetGroupId]['blocked_issue_samples'][] = [
-                    'key' => $issue->key,
-                    'summary' => $issue->summary,
-                    'url' => route('issues.show', ['project' => $project, 'issue' => $issue]),
-                ];
-            }
-
-            foreach ($activeBlockers as $link) {
-                $blockerMeta = $this->resolveBlockerGroupMeta(
-                    $project,
-                    $groups,
-                    $groupIdByIssueId,
-                    $groupBy,
-                    $link->from
-                );
-
-                $blockerGroupId = $blockerMeta['id'];
-
-                if (! isset($groups[$targetGroupId]['_blocked_by'][$blockerGroupId])) {
-                    $groups[$targetGroupId]['_blocked_by'][$blockerGroupId] = [
-                        'external' => $blockerMeta['external'],
-                        'issue_ids' => [],
-                        'label' => $blockerMeta['label'],
-                        'url' => $blockerMeta['url'],
-                    ];
-                }
-
-                $groups[$targetGroupId]['_blocked_by'][$blockerGroupId]['issue_ids'][$issueId] = true;
-
-                if ($blockerGroupId === $targetGroupId) {
-                    $groups[$targetGroupId]['_self_blocked_issue_ids'][$issueId] = true;
-
-                    continue;
-                }
-
-                $edgeKey = $blockerGroupId.'>'.$targetGroupId;
-
-                if (! isset($dependencyEdges[$edgeKey])) {
-                    $dependencyEdges[$edgeKey] = [
-                        'from' => $blockerMeta['label'],
-                        'from_url' => $blockerMeta['url'],
-                        'issue_ids' => [],
-                        'to' => $groups[$targetGroupId]['name'],
-                        'to_url' => $groups[$targetGroupId]['url'],
-                    ];
-                }
-
-                $dependencyEdges[$edgeKey]['issue_ids'][$issueId] = true;
-            }
-        }
-
-        $dependencyEdges = collect($dependencyEdges)
-            ->map(function (array $edge): array {
-                return [
-                    'count' => count($edge['issue_ids']),
-                    'from' => $edge['from'],
-                    'from_url' => $edge['from_url'],
-                    'to' => $edge['to'],
-                    'to_url' => $edge['to_url'],
-                ];
-            })
-            ->sortByDesc('count')
-            ->take(10)
-            ->values()
-            ->all();
-
-        return [$groups, $dependencyEdges];
-    }
-
-    /**
-     * @param  array<string, array<string, mixed>>  $groups
-     * @param  array<string, string>  $groupIdByIssueId
-     * @return array{id:string,label:string,url:?string,external:bool}
-     */
-    private function resolveBlockerGroupMeta(
-        Project $project,
-        array $groups,
-        array $groupIdByIssueId,
-        string $groupBy,
-        Issue $blocker
-    ): array {
-        if ((string) $blocker->project_id === (string) $project->id) {
-            $groupId = $groupIdByIssueId[(string) $blocker->getKey()]
-                ?? ($groupBy === 'parent' ? 'parent:unscoped' : 'milestone:unscheduled');
-
-            return [
-                'id' => $groupId,
-                'label' => $groups[$groupId]['name'] ?? $blocker->key,
-                'url' => $groups[$groupId]['url'] ?? route('issues.show', ['project' => $project, 'issue' => $blocker]),
-                'external' => false,
-            ];
-        }
-
-        $projectKey = $blocker->project?->key ?? 'External';
-
-        if ($groupBy === 'milestone' && $blocker->milestone) {
-            $label = $projectKey.' · '.$blocker->milestone->name;
-
-            return [
-                'id' => 'external:milestone:'.$blocker->milestone->id,
-                'label' => $label,
-                'url' => route('projects.milestones.show', [$blocker->project, $blocker->milestone]),
-                'external' => true,
-            ];
-        }
-
-        return [
-            'id' => 'external:issue:'.$blocker->getKey(),
-            'label' => $projectKey.' · '.$blocker->key,
-            'url' => $blocker->project
-                ? route('issues.show', ['project' => $blocker->project, 'issue' => $blocker])
-                : null,
-            'external' => true,
-        ];
-    }
-
-    /**
-     * @param  array<string, array<string, mixed>>  $groups
      * @param  Collection<string, Issue>  $issuesById
      * @return array<int, array<string, mixed>>
      */
@@ -491,103 +376,104 @@ final class RoadmapService
     {
         $now = now();
 
-        $groups = collect($groups)->map(function (array $group) use ($issuesById, $now): array {
-            $issueIds = array_keys($group['_issue_ids']);
-            $groupIssues = collect($issueIds)
-                ->map(static fn (string $issueId) => $issuesById->get($issueId))
-                ->filter();
+        return collect($groups)
+            ->map(function (array $group) use ($issuesById, $now): array {
+                $issueIds = array_keys($group['_issue_ids']);
+                $groupIssues = collect($issueIds)
+                    ->map(static fn (string $issueId) => $issuesById->get($issueId))
+                    ->filter();
 
-            $doneIssues = $groupIssues->filter(static fn (Issue $issue) => (bool) ($issue->status?->is_done ?? false));
-            $openIssues = $groupIssues->reject(static fn (Issue $issue) => (bool) ($issue->status?->is_done ?? false));
+                $doneIssues = $groupIssues->filter(static fn (Issue $issue) => (bool) ($issue->status?->is_done ?? false));
+                $openIssues = $groupIssues->reject(static fn (Issue $issue) => (bool) ($issue->status?->is_done ?? false));
 
-            $startsAt = $this->resolveWindowStart($groupIssues, $group['starts_at']);
-            $endsAt = $this->resolveWindowEnd($groupIssues, $group['ends_at'], $startsAt);
+                $startsAt = $this->windowResolver->resolveWindowStart($groupIssues, $group['starts_at']);
+                $endsAt = $this->windowResolver->resolveWindowEnd($groupIssues, $group['ends_at'], $startsAt);
 
-            $totalIssues = $groupIssues->count();
-            $doneCount = $doneIssues->count();
-            $openCount = $openIssues->count();
-            $progressPercent = $totalIssues > 0 ? (int) round(($doneCount / $totalIssues) * 100) : 0;
-            $storyPointsTotal = (int) $groupIssues->sum('story_points');
-            $storyPointsDone = (int) $doneIssues->sum('story_points');
-            $blockedIssueCount = count($group['_blocked_issue_ids']);
-            $selfBlockedCount = count($group['_self_blocked_issue_ids']);
-            $overdueIssueCount = $openIssues
-                ->filter(static fn (Issue $issue) => $issue->due_at !== null && $issue->due_at->lt($now))
-                ->count();
-            $unassignedIssueCount = $openIssues
-                ->filter(static fn (Issue $issue) => blank($issue->assignee_id))
-                ->count();
-            $unscheduledIssueCount = $openIssues
-                ->filter(static fn (Issue $issue) => $issue->starts_at === null || $issue->due_at === null)
-                ->count();
+                $totalIssues = $groupIssues->count();
+                $doneCount = $doneIssues->count();
+                $openCount = $openIssues->count();
+                $progressPercent = $totalIssues > 0 ? (int) round(($doneCount / $totalIssues) * 100) : 0;
+                $storyPointsTotal = (int) $groupIssues->sum('story_points');
+                $storyPointsDone = (int) $doneIssues->sum('story_points');
+                $blockedIssueCount = count($group['_blocked_issue_ids']);
+                $selfBlockedCount = count($group['_self_blocked_issue_ids']);
+                $overdueIssueCount = $openIssues
+                    ->filter(static fn (Issue $issue) => $issue->due_at !== null && $issue->due_at->lt($now))
+                    ->count();
+                $unassignedIssueCount = $openIssues
+                    ->filter(static fn (Issue $issue) => blank($issue->assignee_id))
+                    ->count();
+                $unscheduledIssueCount = $openIssues
+                    ->filter(static fn (Issue $issue) => $issue->starts_at === null || $issue->due_at === null)
+                    ->count();
 
-            $readinessPercent = $this->readinessPercent(
-                $progressPercent,
-                $blockedIssueCount,
-                $overdueIssueCount,
-                $unassignedIssueCount,
-                $unscheduledIssueCount,
-                $endsAt
-            );
+                $readinessPercent = $this->readinessPercent(
+                    progressPercent: $progressPercent,
+                    blockedIssueCount: $blockedIssueCount,
+                    overdueIssueCount: $overdueIssueCount,
+                    unassignedIssueCount: $unassignedIssueCount,
+                    unscheduledIssueCount: $unscheduledIssueCount,
+                    endsAt: $endsAt,
+                );
 
-            ['label' => $riskLabel, 'tone' => $riskTone] = $this->riskState(
-                $totalIssues,
-                $progressPercent,
-                $blockedIssueCount,
-                $overdueIssueCount,
-                $unassignedIssueCount,
-                $unscheduledIssueCount,
-                $endsAt
-            );
+                ['label' => $riskLabel, 'tone' => $riskTone] = $this->riskState(
+                    totalIssues: $totalIssues,
+                    progressPercent: $progressPercent,
+                    blockedIssueCount: $blockedIssueCount,
+                    overdueIssueCount: $overdueIssueCount,
+                    unassignedIssueCount: $unassignedIssueCount,
+                    unscheduledIssueCount: $unscheduledIssueCount,
+                    endsAt: $endsAt,
+                );
 
-            $blockedBy = collect($group['_blocked_by'])
-                ->map(static function (array $entry): array {
-                    return [
-                        'count' => count($entry['issue_ids']),
-                        'external' => $entry['external'],
-                        'label' => $entry['label'],
-                        'url' => $entry['url'],
-                    ];
-                })
-                ->sortByDesc('count')
-                ->take(4)
-                ->values()
-                ->all();
+                $blockedBy = collect($group['_blocked_by'])
+                    ->map(static function (array $entry): array {
+                        return [
+                            'count' => count($entry['issue_ids']),
+                            'external' => $entry['external'],
+                            'label' => $entry['label'],
+                            'url' => $entry['url'],
+                        ];
+                    })
+                    ->sortByDesc('count')
+                    ->take(4)
+                    ->values()
+                    ->all();
 
-            return [
-                'id' => $group['id'],
-                'kind' => $group['kind'],
-                'type' => $group['type'],
-                'name' => $group['name'],
-                'subtitle' => $group['subtitle'],
-                'state_label' => $group['state_label'],
-                'url' => $group['url'],
-                'starts_at' => $startsAt,
-                'ends_at' => $endsAt,
-                'date_window_label' => $this->formatWindow($startsAt, $endsAt),
-                'total_issues' => $totalIssues,
-                'done_issues' => $doneCount,
-                'open_issues' => $openCount,
-                'story_points_total' => $storyPointsTotal,
-                'story_points_done' => $storyPointsDone,
-                'progress_percent' => $progressPercent,
-                'readiness_percent' => $readinessPercent,
-                'blocked_issue_count' => $blockedIssueCount,
-                'self_blocked_issue_count' => $selfBlockedCount,
-                'overdue_issue_count' => $overdueIssueCount,
-                'unassigned_issue_count' => $unassignedIssueCount,
-                'unscheduled_issue_count' => $unscheduledIssueCount,
-                'scope_count' => count($group['_scope_ids']),
-                'milestone_count' => count($group['_milestone_ids']),
-                'risk_label' => $riskLabel,
-                'risk_tone' => $riskTone,
-                'risk_badge_class' => $this->badgeClassForTone($riskTone),
-                'chart_color' => $this->chartColorForTone($riskTone),
-                'type_badge_class' => $this->typeBadgeClass($group['type']),
-                'blocked_by' => $blockedBy,
-                'blocked_issue_samples' => $group['blocked_issue_samples'],
-            ];
-        })
+                return [
+                    'id' => $group['id'],
+                    'kind' => $group['kind'],
+                    'type' => $group['type'],
+                    'name' => $group['name'],
+                    'subtitle' => $group['subtitle'],
+                    'state_label' => $group['state_label'],
+                    'url' => $group['url'],
+                    'starts_at' => $startsAt,
+                    'ends_at' => $endsAt,
+                    'date_window_label' => $this->windowResolver->formatWindow($startsAt, $endsAt),
+                    'total_issues' => $totalIssues,
+                    'done_issues' => $doneCount,
+                    'open_issues' => $openCount,
+                    'story_points_total' => $storyPointsTotal,
+                    'story_points_done' => $storyPointsDone,
+                    'progress_percent' => $progressPercent,
+                    'readiness_percent' => $readinessPercent,
+                    'blocked_issue_count' => $blockedIssueCount,
+                    'self_blocked_issue_count' => $selfBlockedCount,
+                    'overdue_issue_count' => $overdueIssueCount,
+                    'unassigned_issue_count' => $unassignedIssueCount,
+                    'unscheduled_issue_count' => $unscheduledIssueCount,
+                    'scope_count' => count($group['_scope_ids']),
+                    'milestone_count' => count($group['_milestone_ids']),
+                    'risk_label' => $riskLabel,
+                    'risk_tone' => $riskTone,
+                    'risk_badge_class' => $this->badgeClassForTone($riskTone),
+                    'chart_color' => $this->chartColorForTone($riskTone),
+                    'type_badge_class' => $this->typeBadgeClass($group['type']),
+                    'blocked_by' => $blockedBy,
+                    'blocked_issue_samples' => $group['blocked_issue_samples'],
+                ];
+            })
             ->sortBy(function (array $group): string {
                 $synthetic = in_array($group['type'], ['unscheduled', 'unscoped'], true) ? 1 : 0;
                 $startsAt = $group['starts_at']?->getTimestamp() ?? 9999999999;
@@ -597,63 +483,6 @@ final class RoadmapService
             })
             ->values()
             ->all();
-
-        return $groups;
-    }
-
-    /**
-     * @param  Collection<int, Issue>  $groupIssues
-     */
-    private function resolveWindowStart(Collection $groupIssues, ?Carbon $startsAt): ?Carbon
-    {
-        if ($startsAt !== null) {
-            return $startsAt->copy();
-        }
-
-        $candidate = $groupIssues
-            ->flatMap(static fn (Issue $issue): array => [
-                $issue->starts_at,
-                $issue->created_at,
-                $issue->milestone?->starts_at,
-            ])
-            ->filter()
-            ->map(fn ($value) => $this->normalizeDate($value))
-            ->sortBy(static fn (Carbon $date) => $date->getTimestamp())
-            ->first();
-
-        return $candidate?->copy();
-    }
-
-    /**
-     * @param  Collection<int, Issue>  $groupIssues
-     */
-    private function resolveWindowEnd(Collection $groupIssues, ?Carbon $endsAt, ?Carbon $startsAt): ?Carbon
-    {
-        $candidate = $endsAt?->copy();
-
-        if ($candidate === null) {
-            $candidate = $groupIssues
-                ->flatMap(static fn (Issue $issue): array => [
-                    $issue->due_at,
-                    $issue->updated_at,
-                    $issue->milestone?->released_at,
-                    $issue->milestone?->due_at,
-                ])
-                ->filter()
-                ->map(fn ($value) => $this->normalizeDate($value))
-                ->sortByDesc(static fn (Carbon $date) => $date->getTimestamp())
-                ->first();
-        }
-
-        if ($candidate === null && $startsAt !== null) {
-            $candidate = $startsAt->copy()->addDays(7);
-        }
-
-        if ($candidate !== null && $startsAt !== null && $candidate->lt($startsAt)) {
-            $candidate = $startsAt->copy()->addDay();
-        }
-
-        return $candidate?->copy();
     }
 
     private function readinessPercent(
@@ -729,7 +558,10 @@ final class RoadmapService
         array $rootScopeIds,
         array $groups
     ): array {
-        $releaseCount = $milestones->filter(fn (Milestone $milestone) => $this->milestoneType($milestone) === 'release')->count();
+        $releaseCount = $milestones
+            ->filter(fn (Milestone $milestone) => $this->milestoneType($milestone) === 'release')
+            ->count();
+
         $upcomingRelease = $milestones
             ->filter(fn (Milestone $milestone) => $this->milestoneType($milestone) === 'release')
             ->sortBy(static function (Milestone $milestone): int {
@@ -764,9 +596,9 @@ final class RoadmapService
             'standalone_issue_count' => $standaloneIssueCount,
             'upcoming_release' => $upcomingRelease ? [
                 'label' => trim($upcomingRelease->name.' '.($upcomingRelease->version ?? '')),
-                'window' => $this->formatWindow(
-                    $this->normalizeDate($upcomingRelease->starts_at),
-                    $this->normalizeDate($upcomingRelease->released_at ?? $upcomingRelease->due_at)
+                'window' => $this->windowResolver->formatWindow(
+                    $this->windowResolver->normalizeDate($upcomingRelease->starts_at),
+                    $this->windowResolver->normalizeDate($upcomingRelease->released_at ?? $upcomingRelease->due_at),
                 ),
             ] : null,
         ];
@@ -833,14 +665,10 @@ final class RoadmapService
 
     /**
      * @param  Collection<int, Milestone>  $milestones
-     * @param  array<int, array{id:string,label:string}>  $milestoneOptions
      */
-    private function resolveMilestoneSelection(
-        Collection $milestones,
-        array $milestoneOptions,
-        ?string $selectedMilestoneId
-    ): ?string {
-        if ($selectedMilestoneId && collect($milestoneOptions)->contains(static fn (array $option): bool => $option['id'] === $selectedMilestoneId)) {
+    private function resolveMilestoneSelection(Collection $milestones, ?string $selectedMilestoneId): ?string
+    {
+        if ($selectedMilestoneId && $milestones->contains(static fn (Milestone $milestone): bool => (string) $milestone->id === $selectedMilestoneId)) {
             return $selectedMilestoneId;
         }
 
@@ -852,182 +680,6 @@ final class RoadmapService
             ->first();
 
         return $releaseSelection?->id ?? $milestones->first()?->id;
-    }
-
-    /**
-     * @param  Collection<int, Issue>  $issues
-     * @return array<string, mixed>
-     */
-    private function buildBurnChart(?Milestone $milestone, Collection $issues): array
-    {
-        if (! $milestone instanceof Milestone) {
-            return [
-                'empty' => true,
-                'labels' => [],
-                'series' => [],
-                'stats' => [
-                    'done' => 0,
-                    'open' => 0,
-                    'total' => 0,
-                ],
-                'subtitle' => null,
-                'title' => 'Burnup / burndown',
-                'window' => null,
-            ];
-        }
-
-        $doneStatusIds = IssueStatus::query()
-            ->where('is_done', true)
-            ->pluck('id')
-            ->map(static fn ($statusId): int => (int) $statusId)
-            ->all();
-
-        $eventsByIssueId = $issues->isEmpty()
-            ? collect()
-            : IssueStatusEvent::query()
-                ->whereIn('issue_id', $issues->modelKeys())
-                ->orderBy('changed_at')
-                ->get(['issue_id', 'to_status_id', 'changed_at'])
-                ->groupBy('issue_id');
-
-        $doneAtByIssueId = [];
-        $latestDoneAt = null;
-
-        foreach ($issues as $issue) {
-            $doneEvent = $eventsByIssueId
-                ->get((string) $issue->getKey(), collect())
-                ->first(static fn ($event): bool => in_array((int) $event->to_status_id, $doneStatusIds, true));
-
-            $doneAt = $doneEvent?->changed_at
-                ? $this->normalizeDate($doneEvent->changed_at)
-                : (($issue->status?->is_done ?? false) ? $this->normalizeDate($issue->updated_at) : null);
-
-            $doneAtByIssueId[(string) $issue->getKey()] = $doneAt;
-
-            if ($doneAt && ($latestDoneAt === null || $doneAt->gt($latestDoneAt))) {
-                $latestDoneAt = $doneAt;
-            }
-        }
-
-        $start = $this->normalizeDate($milestone->starts_at)
-            ?? $issues
-                ->map(fn (Issue $issue) => $this->normalizeDate($issue->created_at))
-                ->filter()
-                ->sortBy(static fn (Carbon $date) => $date->getTimestamp())
-                ->first()
-            ?? now()->copy()->startOfDay();
-
-        $endCandidates = collect([
-            $this->normalizeDate($milestone->released_at),
-            $this->normalizeDate($milestone->due_at),
-            $latestDoneAt,
-            $issues
-                ->map(fn (Issue $issue) => $this->normalizeDate($issue->created_at))
-                ->filter()
-                ->sortByDesc(static fn (Carbon $date) => $date->getTimestamp())
-                ->first(),
-            now()->copy()->startOfDay(),
-        ])->filter();
-
-        $end = $endCandidates
-            ->sortByDesc(static fn (Carbon $date) => $date->getTimestamp())
-            ->first()
-            ?->copy() ?? $start->copy()->addDays(14);
-
-        if ($end->lt($start)) {
-            $end = $start->copy()->addDays(14);
-        }
-
-        $scopeStarts = [];
-        $doneDates = [];
-
-        foreach ($issues as $issue) {
-            $scopeStart = $this->normalizeDate($issue->created_at)?->startOfDay() ?? $start->copy();
-            if ($scopeStart->lt($start)) {
-                $scopeStart = $start->copy();
-            }
-
-            $scopeStarts[$scopeStart->toDateString()] = ($scopeStarts[$scopeStart->toDateString()] ?? 0) + 1;
-
-            $doneAt = $doneAtByIssueId[(string) $issue->getKey()] ?? null;
-            if (! $doneAt instanceof Carbon) {
-                continue;
-            }
-
-            $doneDate = $doneAt->copy()->startOfDay();
-            if ($doneDate->lt($start)) {
-                $doneDate = $start->copy();
-            }
-
-            $doneDates[$doneDate->toDateString()] = ($doneDates[$doneDate->toDateString()] ?? 0) + 1;
-        }
-
-        $labels = [];
-        $scopeSeries = [];
-        $doneSeries = [];
-        $remainingSeries = [];
-        $scope = 0;
-        $done = 0;
-        $cursor = $start->copy();
-
-        while ($cursor->lte($end)) {
-            $key = $cursor->toDateString();
-            $labels[] = $cursor->format('M j');
-
-            $scope += (int) ($scopeStarts[$key] ?? 0);
-            $done += (int) ($doneDates[$key] ?? 0);
-
-            $scopeSeries[] = $scope;
-            $doneSeries[] = $done;
-            $remainingSeries[] = max(0, $scope - $done);
-
-            $cursor->addDay();
-        }
-
-        return [
-            'empty' => false,
-            'labels' => $labels,
-            'note' => 'Scope uses issue creation dates and first done transitions within the milestone.',
-            'series' => [
-                ['name' => 'Scope', 'data' => $scopeSeries],
-                ['name' => 'Done', 'data' => $doneSeries],
-                ['name' => 'Remaining', 'data' => $remainingSeries],
-            ],
-            'stats' => [
-                'done' => $issues->filter(static fn (Issue $issue) => (bool) ($issue->status?->is_done ?? false))->count(),
-                'open' => $issues->filter(static fn (Issue $issue) => ! ($issue->status?->is_done ?? false))->count(),
-                'total' => $issues->count(),
-            ],
-            'subtitle' => $milestone->version,
-            'title' => $milestone->name,
-            'window' => $this->formatWindow($start, $end),
-        ];
-    }
-
-    private function normalizeDate(mixed $value): ?Carbon
-    {
-        if ($value === null) {
-            return null;
-        }
-
-        return Carbon::parse($value);
-    }
-
-    private function formatWindow(?Carbon $startsAt, ?Carbon $endsAt): ?string
-    {
-        if ($startsAt && $endsAt) {
-            return $startsAt->format('M j').' to '.$endsAt->format('M j');
-        }
-
-        if ($startsAt) {
-            return 'Starts '.$startsAt->format('M j');
-        }
-
-        if ($endsAt) {
-            return 'Targets '.$endsAt->format('M j');
-        }
-
-        return null;
     }
 
     private function milestoneType(Milestone $milestone): string
